+++ install/js/mb/admin/kit/config.php
@@ -2,10 +2,18 @@
 defined('B_PROLOG_INCLUDED') || die();
 
 return [
-    'js' => 'dist/kit.bundle.js',
-    'css' => 'dist/kit.bundle.css',
+    'js' => [
+        'dist/kit.bundle.js'
+    ],
+    'css' => [
+        'dist/kit.bundle.css'
+    ],
     'rel' => [
-		'main.polyfill.core',
+		'main.core.collections',
+		'main.core.events',
+		'main.loader',
+		'ui.entity-selector',
+		'main.core',
 	],
-    'skip_core' => true,
+	'skip_core' => false,
 ];