+++ install/js/mb/ui/tabs/bundle.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-	input: 'src/index.js',
-	output: 'dist/tabs.bundle.js',
-	namespace: 'MB.UI.Tabs',
-	cssImages: {
-		type: 'copy',
-	},
-};